move linker file into graph
