+++ internal/bundler/bundler.go
@@ -32,14 +32,6 @@ import (
 	"github.com/evanw/esbuild/internal/xxhash"
 )
 
-type EntryPointKind uint8
-
-const (
-	EntryPointNone EntryPointKind = iota
-	EntryPointUserSpecified
-	EntryPointDynamicImport
-)
-
 type scannerFile struct {
 	module     graph.Module
 	pluginData interface{}
@@ -48,36 +40,6 @@ type scannerFile struct {
 	// about this file in JSON format. This is a partial JSON file that will be
 	// fully assembled later.
 	jsonMetadataChunk string
-}
-
-type LinkerFile struct {
-	Module graph.Module
-
-	// The minimum number of links in the module graph to get from an entry point
-	// to this file
-	DistanceFromEntryPoint uint32
-
-	// This is true if this file has been marked as live by the tree shaking
-	// algorithm.
-	IsLive bool
-
-	// This holds all entry points that can reach this file. It will be used to
-	// assign the parts in this file to a chunk.
-	EntryBits helpers.BitSet
-
-	// If "entryPointKind" is not "entryPointNone", this is the index of the
-	// corresponding entry point chunk.
-	EntryPointChunkIndex uint32
-
-	// This file is an entry point if and only if this is not "entryPointNone".
-	// Note that dynamically-imported files are allowed to also be specified by
-	// the user as top-level entry points, so some dynamically-imported files
-	// may be "entryPointUserSpecified" instead of "entryPointDynamicImport".
-	EntryPointKind EntryPointKind
-}
-
-func (f *LinkerFile) IsEntryPoint() bool {
-	return f.EntryPointKind != EntryPointNone
 }
 
 // This is data related to source maps. It's computed in parallel with linking
@@ -1879,7 +1841,7 @@ func (b *Bundle) Compile(log logger.Log, options config.Options) ([]graph.Output
 		options.OutputFormat = config.FormatESModule
 	}
 
-	files := make([]LinkerFile, len(b.files))
+	files := make([]graph.LinkerFile, len(b.files))
 	for i, file := range b.files {
 		files[i].Module = file.module
 	}
@@ -1987,7 +1949,7 @@ func (b *Bundle) Compile(log logger.Log, options config.Options) ([]graph.Output
 // deterministic given that the entry point order is deterministic, since the
 // returned order is the postorder of the graph traversal and import record
 // order within a given file is deterministic.
-func findReachableFiles(files []LinkerFile, entryPoints []entryMeta) []uint32 {
+func findReachableFiles(files []graph.LinkerFile, entryPoints []entryMeta) []uint32 {
 	visited := make(map[uint32]bool)
 	var order []uint32
 	var visit func(uint32)
